zoom/requirements.txt
# Supabase
supabase>=2.0.0

# Google Gemini
google-generativeai>=0.3.0

# Environment
python-dotenv>=1.0.0

# Google Sheets / Drive API
google-auth>=2.0.0
google-api-python-client>=2.0.0
gspread>=5.0.0

# HTTP requests (Zoom API)
requests>=2.28.0
